ML_week_02/end_to_end_pipeline/requirements.txt
# ============================================================
# Production dependencies
# ============================================================
pandas>=2.0,<3.0
numpy>=1.24,<2.0
scikit-learn>=1.3,<2.0
xgboost>=2.0,<3.0
mlflow>=2.9,<3.0
optuna>=3.4,<4.0
fastapi>=0.104,<1.0
uvicorn[standard]>=0.24,<1.0
pydantic>=2.5,<3.0
PyYAML>=6.0,<7.0
joblib>=1.3,<2.0
matplotlib>=3.8,<4.0
httpx>=0.25,<1.0
